requirements.txt
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
psycopg[binary]
psycopg-pool
python-dotenv
